refactoring to use jdk 1.5
git-svn-id: svn://localhost/trunk@568 f2a0fb1d-5e3f-4af0-b388-fe59793aa704

+++ src/java/ironfist/ReverseIntegerComparator.java
@@ -4,15 +4,15 @@ import java.util.Comparator;
 
 /**
  * @author pmgremo
- *  
+ * 
  */
 public class ReverseIntegerComparator implements Comparator {
 
-  /**
-   * @see java.util.Comparator#compare(Object, Object)
-   */
-  public int compare(Object o1, Object o2) {
-    return -(((Integer) o1).compareTo(o2));
-  }
+	/**
+	 * @see java.util.Comparator#compare(Object, Object)
+	 */
+	public int compare(Object o1, Object o2) {
+		return -(((Integer) o1).compareTo((Integer) o2));
+	}
 
 }
